smart `C-h/j/k/l` switching between vim or tmux panes
It works by detecting if Vim is running in the current pane and
forwarding `C-h/j/k/l` to it; otherwise it simply switches tmux panes as
usual. If Vim accepted the keystroke, it should decide whether it should
switch windows in the given direction or pass the `select-pane` command
back to tmux if needed.