Add `travis-log`, alternative to `travis logs`
Bypasses the official (slow) `travis` CLI tool and accesses the API
directly to retrieve all logs for the last completed build of the
current branch.

Caches the result, so that subsequent lookups for the same SHA are
instantaneous.

When displaying logs, strip the contents of Travis "folds", which is git
clone output, bundler install output, and similar.